backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
requests==2.32.3
google-cloud-speech==2.21.0
openai==1.12.0
pydub==0.25.1
python-dotenv==1.0.0
whitenoise==6.5.0
gunicorn==21.2.0
pytest==7.4.0
pytest-django==4.5.2
